Hook up publish/subscribe 'properly'.
This should greatly reduce the amount of update traffic the client
requests/gets when they are not on the main blackboard page.
(Also, it limits chat updates to the current chat room.)

Added subscriptions that can be used for the paged chat log and
operation log archive pages... when I implement those.