Fix callbacks on initial visit to chat page.
Presence information wasn't being sent because the callback which invokes
joinRoom wasn't being invoked after the "pick a nick" modal was hidden.
Turns out that the modal was being torn down and destroyed by meteor
when we frobbed the 'nickModalVisible' session state, causing all the
events registered on it to disappear.  Use Template.preserve to ensure
that the #nickPickModal <div> is reused (preserving its events) instead
of being regenerated from scratch.

Also, actually invoke the callback, instead of just returning it as a value.