Don't keep reattaching localScroll events.  Fixes #3.
Every time the reactivity context changed for the blackboard template, it
would rerun the localScroll plugin.  For most jquery plugins this is harmless;
attaching a behavior to a node which already has one is either ignored or
updates the properties of the behavior.  But localScroll was adding new
event handlers each time.  The blackboard page does about 60 'reactivity
updates' per minute, so the overload of event handlers would cause
problems relatively quickly.

Now we only attach the localScroll stuff when the template is first
created -- and we "preserve" the #bb-sidebar to ensure that that DOM
node doesn't get discarded and rebuilt between render runs.