Use Session.equals() wherever possible for more precise dependencies.
Session.equals() introduces a dependency on a specific *value* for a variable,
as opposed to a dependency on the variable in general.  That is, code
dependent on Session.equals('currentPage', 'chat') will only execute if
currentPage is equal to 'chat' (or stops being equal to 'chat'), not
whenever currentPage changes.