Share "last read message" state across tabs; use it for notifications.
The number of unread messages is displayed in the favicon, so you can
see it at a glance.  We also display the "last read message" as a bar
across the chat window, so you can easily see how far back you need
to scroll to catch up.  Remove a bunch of now-unnecessary code.

Closes #45, #94.

Change-Id: Ie393b5e5226ad2c702ee5364d7991f31bb08d9a8