Paginate on server side to avoid the need for `limit` in chat message queries.
Using `limit` in message queries prevents us from using the oplog-tailing
server observe code, and it also complicates the client update efficiency
because we can't use a cursor directly -- we have to reverse the sort order
when using the limit query, and that's incompatible with the optimizations
which the renderer does for lists.

Use an explicit Pages collection, maintained by the server, which splits
the message logs for a given room into roughly MESSAGE_PAGE-sized pieces.
The `limit` queries can then be replaced by simple timestamp queries inside
an interval.

Make the pages into a doubly-linked list, which also allows us to easily
page forward as well as backward through the message history.

Change-Id: I38393c0f2b1df1b2dcb1201d34e7e70b73a4ba19