Try to better manage flicker on chat newMessage.
The old flicker-suppression games made things worse.  We basically want to
try to get only a single update to the display.  If we allow the client
to insert the message w/ the wrong timestamp, we could potentially see
discrete events as the faked message is removed, the correct server message
is added, the faked last read is removed, and the new last read is added.

Instead do the update only on the server. It's better to pause a
little while after the user hits send and before the message shows up.
Also update lastread before inserting the new message so that lastread
appears to disappear before appearing with the new message.  (That's
less distracting then having to add the new message w/o the correct
last read, then updating both the new message and the old message to
redraw the last read marker in the right spot.)

Change-Id: I8c800643218f5a189a82fdf1935d2bb21e89b68d