Add button to toolbar to mute codebot.
Tag messages with a special class if they come from the bot (and aren't
a PM reply to the user), and toggle a class on the document root.  This
is much faster than redoing a query on all the messages, or using spark
to toggle a class, both of which force a rendering from scratch of the
entire list of messages.

Issue #76

Change-Id: I1def40e4b65e1540eef9abce04720723a6b69dd6