Load hubot scripts from codex-blackboard-hubot-scripts repository.
Meteor's module loading isn't really compatible with hubot's package-based
script system.  Create a separate npm module to encapsulate all the
hubot script package dependencies and allow the use of the standard
hubot script loader.  We create a local Meteor package in `packages/`
and reference the npm module version using a git hash to avoid the need to
publish to npm and/or atmosphere every time we make a change to the
set of hubot scripts.  Instead we just need to push to github and update
the hash in `packages/codex-blackboard-hubot-scripts/package.js`.

Change-Id: I5f98a05339a4a685850185152d0992d07d5641a3