Use Template.onCreated/onRendered/onDestroyed/autorun/subscribe
These methods were renamed in 1.0.4.2, and this.subscribe was added
to automatically unsubscribe when the template is destroyed.  This
new pattern should be more robust against subscription leaks; I've
tried to minimize the number of "bare" calls to Meteor.subscribe
(which are the only places that could possibly leak now).

Change-Id: I7bad63b80a2e1b4782d81f10c3c658e7882d078a