Add `OldMessages` collection for archived messages.
Try to keep the `Messages` collection small, since that's what most
people will be loading/watching, and Meteor seems to still do linear
passes over it from time to time.

This code adds the OldMessages collection, but doesn't move anything
from Messages to OldMessages (that will be in a follow-up patch).

Change-Id: I3a0686af8452e50e0e5c15be3ab2c0ee5769381b