Enable desktop notifications from Oplog messages
* Add stream field to oplogs for notification deduping/muting.
Also made it part of share.model and set the field on existing calls.

* Notification stream for stuck puzzles as well.

* Add "bot announce" command to add to oplog.

* Add answer to oplog for setAnswer.
Previously it was only in the created/canceled callins, but with
notifications, someone might subscribe to answers but not callins.

* Use JS Notification API to add desktop notifications for oplog events.

* Add UI to modify which notifications are listened for.
Also changes renotify and tag, as the intent was to both dedupe and
rate-limit, but at least on mac, a second notification for the same
stream won't appear if you didn't manually close the first, which is
probably hidden in the notifications menu you'd never check.

* Button to click before asking for notification permission.

Only appears if notification permission is in default state.
Disappears after click if permission is granted or refused.

* Fake service worker to allow notifications on android Chrome.

Unfortunately we can't compile meteor code to a real service worker
javascript file, which means we can't generate notifications while
the screen is off, though they will be picked up as soon as the screen
is turned on for any reason.
(Service worker support likely coming in 0.6.2; see
https://github.com/NitroBAY/meteor-service-worker
https://github.com/meteor/meteor-feature-requests/issues/18
https://github.com/meteor/meteor-feature-requests/issues/19
https://github.com/meteor/meteor/pull/9439
https://github.com/meteor/meteor/milestone/47 )

This feature written by Dan Rosart, hacked to piece by C. Scott Ananian
(tweaks to UX for compactness, etc).