Use IntersectionObserver to handle chat "scroll to bottom" functionality
The legacy browser visibility APIs are hideous, and we wrestled with lots
of weird corner cases trying to determine if the user was actually scrolled
to the bottom of the page using the scrollTop and related properties.

The new(ish) IntersectionObserver API gives us a simple straightforward
way to ask the question we really want to know: are we scrolled to the
bottom of the page or not.  Leave the old buggy code in place for now
for legacy browsers (sigh).